requirements.txt
httpx
python-dotenv
pydantic
fastapi
uvicorn
jinja2
python-multipart
